improvement: convert Shiritori commands to CLI
Music commands were renamed to improve clarity and avoid name conflicts.
