+++ src/main/kotlin/com/github/frozensync/discord/cli/YggdrasilCli.kt
@@ -2,9 +2,10 @@ package com.github.frozensync.discord.cli
 
 import com.github.ajalt.clikt.core.subcommands
 import com.github.frozensync.*
-import com.github.frozensync.music.JoinCommand
+import com.github.frozensync.games.shiritori.*
+import com.github.frozensync.music.JoinVoiceChannelCommand
 import com.github.frozensync.music.Music
-import com.github.frozensync.music.PlayCommand
+import com.github.frozensync.music.PlayMusicCommand
 import discord4j.core.event.domain.message.MessageCreateEvent
 
 internal class Yggdrasil : AbstractCommand(name = "@Yggdrasil") {
@@ -18,13 +19,20 @@ internal class Yggdrasil : AbstractCommand(name = "@Yggdrasil") {
     }
 }
 
-internal val yggdrasilCli : () -> Yggdrasil = {
+internal val yggdrasilCli: () -> Yggdrasil = {
     Yggdrasil()
         .subcommands(
+            ShiritoriCategory()
+                .subcommands(
+                    NewGameCommand(),
+                    PlayWordCommand(),
+                    ForfeitCommand(),
+                    CurrentTurnCommand()
+                ),
             Music()
                 .subcommands(
-                    JoinCommand(),
-                    PlayCommand(),
+                    JoinVoiceChannelCommand(),
+                    PlayMusicCommand(),
                 ),
             HealthCheck()
                 .subcommands(
